Restructure Solid UI: src/components + src/App.jsx + bridges in src/setup
- Rename src/app/ tree to src/components/ (feature-scoped JSX)
- Move App.jsx + mount-app.jsx up to src/
- Move UI bridges (setupAppRoot, setupContactsAppBusHandlers) to src/setup/
- Update eslint boundaries: rename 'app' element to 'components';
  components allowed to import components/auth/feature/shared; setup may
  import components (not the other way around)
- Update WIP architecture docs to match new layout
