+++ docs/WIP_Architecture/LINT_ENFORCEMENT.md
@@ -10,7 +10,7 @@ What is automatically enforced today. Source of truth: `eslint.config.js`, `esli
   - Block `**/auth-state`, `**/auth-state.js` imports outside `src/auth/`.
 - `boundaries/dependencies` — per-module import boundaries (incremental rollout).
 - JSX parsing is enabled for `src/**/*.jsx`.
-- Boundary lint includes the `app` layer in addition to `setup`, `auth`, `feature`, and `shared`.
+- Boundary lint includes the `components` layer in addition to `setup`, `auth`, `feature`, and `shared`.
 
 ## Current status (April 15, 2026)
 
@@ -30,9 +30,9 @@ When `contacts-state.js`, `messaging-state.js`, or `call-state.js` lands, extend
 
 ## Current Solid coverage
 
-- `src/app/**` is linted as the `app` layer.
-- `app` may import `app`, `auth`, `feature`, `shared`, and `setup`.
-- `setup` may import `app` so bootstrap code can mount and bridge Solid UI.
+- `src/components/**` is linted as the `components` layer.
+- `components` may import `components`, `auth`, `feature`, and `shared`.
+- `setup` may import `components` so bootstrap code can bridge Solid UI.
 
 ---
