+++ src/setup/setupAppRoot.js
@@ -5,8 +5,8 @@
 import { produce } from 'solid-js/store';
 import { subscribe, dispatchCommand } from '../shared/events/index.js';
 import { getAllContactsSorted } from '../features/contacts/index.js';
-import { setupContactsAppBusHandlers } from './contacts/contacts-command-handlers.js';
-import { setContacts } from './contacts/ContactsList.jsx';
+import { setupContactsAppBusHandlers } from './setupContactsAppBusHandlers.js';
+import { setContacts } from '../components/contacts/ContactsList.jsx';
 
 /**
  * @typedef {Object} ContactRow
